pyproject.toml
[project]
name = "agent-arena"
version = "0.1.0"
description = "AI Agent Society Emulator - Multiple AI agents interacting in an IRC-like chat environment"
requires-python = ">=3.11"
dependencies = [
    "claude-agent-sdk>=0.1.0",
    "anthropic>=0.40.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sqlalchemy>=2.0.0",
    "click>=8.1.0",
    "rich>=13.0.0",
    "pyyaml>=6.0.0",
    "python-dotenv>=1.0.0",
    "jinja2>=3.1.0",
    "websockets>=13.0.0",
    "aiosqlite>=0.20.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
]

[project.scripts]
arena = "agent_arena.cli.main:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["agent_arena"]
